Updates after backend changes, missing few things

+++ lib/ui/widgets/creators/stats_box_row.dart
@@ -2,8 +2,8 @@ import 'package:d_reader_flutter/ui/widgets/creators/stats_box.dart';
 import 'package:flutter/material.dart';
 
 class StatsBoxRow extends StatelessWidget {
-  final double totalVolume;
-  final double issuesCount;
+  final int totalVolume;
+  final int issuesCount;
   const StatsBoxRow({
     Key? key,
     required this.totalVolume,
